==> 00_Inicio.py <==
import streamlit as st

st.session_state["reserva_seleccionada"] = None
st.session_state["continuar"] = False
st.session_state.paso = 0

st.markdown("""
# 🚘 Plataforma de Alquiler de Vehículos

Bienvenido a tu plataforma confiable y rápida para alquilar vehículos.  
Ya sea que necesites un auto ¡estás en el lugar correcto!

Usá el menú lateral para:

- 📝 **Registrarte** como nuevo usuario  
- 🔐 **Iniciar sesión** para acceder a tu cuenta  
- 🚗 **Explorar el catálogo** de vehículos disponibles

Además, podés consultar tus reservas, ver detalles de los vehículos y gestionar tus alquileres de forma simple con solo Iniciar Sesión"

---


¡Comenzá tu viaje hoy mismo!
""")
